+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.6.4] - 2025-11-15
 
-- Set default passphrase to `quit`, so user can skip setup and immediately try out HandsOff
+- Set default passphrase to `qwet`, so user can skip setup and immediately try out HandsOff (single-hand unlock)
 
 ## [0.6.3] - 2025-11-15